+++ mne_bids/tests/test_report.py
@@ -10,7 +10,7 @@ from mne.datasets import testing
 from mne.utils import _TempDir
 
 from mne_bids import (make_bids_basename,
-                      create_methods_paragraph)
+                      make_report)
 from mne_bids.write import write_raw_bids
 
 subject_id = '01'
@@ -41,15 +41,18 @@ def test_read_raw_kind():
     write_raw_bids(raw, bids_basename, bids_root, overwrite=True,
                    verbose=False)
 
-    methods_summ = create_methods_paragraph(bids_root)
-    expected_report = """This dataset was created with BIDS version 1.4.0 by MNE-BIDS. This report was
-generated with MNE-BIDS (https://doi.org/10.21105/joss.01896). There are 1
-subjects. The dataset consists of 1 recording sessions 01. The sex of the
-subjects are all unknown. The handedness of the subjects are all unknown. The
-ages of the subjects are all unknown. Data was acquired using a MEG system
-(Elekta manufacturer) with line noise at 60 Hz. There are 1 total scans, 376.0
-+/- 0.0 total recording channels per scan (374.0 +/- 0.0 are used and 2.0 +/-
-0.0 are removed from analysis). The dataset lengths range from 20.0 to 20.0
-seconds, for a total of 20.0 seconds of data recorded over all scans (20.0 +/-
-0.0)."""  # noqa
-    assert methods_summ == expected_report
+    report = make_report(bids_root)
+    print(report)
+
+    expected_report = \
+    """This dataset was created with BIDS version 1.4.0 by Please cite MNE-BIDS in your
+publication before removing this (citations in README). This report was
+generated with MNE-BIDS (https://doi.org/10.21105/joss.01896). There is 1
+subject. The dataset consists of 1 recording sessions 01.The sex of the subjects
+are all unknown. The handedness of the subjects are all unknown. The ages of the
+subjects are all unknown. Data was acquired using a MEG system (Elekta
+manufacturer) with line noise at 60 Hz . There is 1 scan in total, 376.0 +/- 0.0
+total recording channels per scan (374.0 +/- 0.0 are used and 2.0 +/- 0.0 are
+removed from analysis). The dataset lengths range from 20.0 to 20.0 seconds, for
+a total of 20.0 seconds of data recorded over all scans (20.0 +/- 0.0)."""  # noqa
+    assert report == expected_report